requirements.txt
# HTTP & Data
requests>=2.31.0
pandas>=2.0.0
numpy>=1.24.0

# Environment
python-dotenv>=1.0.0
